Refactor business.service and provision.service to use new api file with config state
