+++ packages/webcomponents/src/api/services/provision.service.ts
@@ -1,4 +1,8 @@
-import Api, { IApiResponse } from '../Api';
+import { IApiResponse } from '../Api';
+import Api from '../ApiNew';
+
+
+const api = Api();
 
 export interface IProvisionService {
   postProvisioning(
@@ -15,13 +19,10 @@ export class ProvisionService implements IProvisionService {
     product: string
   ): Promise<IApiResponse<any>> {
     const endpoint = `entities/provisioning`;
-    const payload = {
+    const body = {
       business_id: businessId,
       product_category: product,
     };
-    return Api({ authToken, apiOrigin: PROXY_API_ORIGIN }).post({
-      endpoint,
-      body: payload,
-    });
+    return api.post({ endpoint, body, authToken });
   }
 }
